Move tests for parallel lambda from tests/ into werkit/ and restore dev.py
